// src/acenda/models/acenda-category.model.ts
export class AcendaCategory {
  parent_id?: number[];
  status?: string;
  track_status?: boolean;
  selection?: string;
  query?: string;
  slug?: string;
  name: string;
  description?: string;
  position?: number;
  facets?: string[];
  image_type?: string;
  image_product_id?: number;
  image_sort?;
  images?;
  page_title?: string;
  meta_description?: string;
  meta_keywords?: string;
  date_created?: Date;
  date_modified?: string;
  thumbnail?: string;
  has_items?: boolean;
  number_products?: number;
  number_categories?: number;
}