// src/acenda/models/acenda-collection.model.ts
export class AcendaCollection {
  acenda_id?: number;
  group?: string;
  status: string;
  slug?: string;
  name: string;
  category_id?: number[];
  customer_group_id?: number[];
  brand?: string;
  tags?: string[];
  description?: string;
  cross_sellers?: number[];
  review_score?: number;
  images?: string[];
  videos?: string[];
  layout?: string[];
  page_title?: string;
  meta_keywords?: string;
  meta_description?: string;
  date_modified?: string;
  date_created?: Date;
  keywords?: string;
  category?: string[];
  title?: string;

}